requirements/base.txt
anyio==3.6.1
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
fastapi==0.85.0
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.10.1
platformdirs==2.5.2
protobuf==4.21.7
pycparser==2.21
pycrypto==2.6.1
pydantic==1.10.2
PyJWT==2.5.0
pytz==2022.4
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.18.3
openpyxl==3.1.0
xmltodict==0.13.0
xlsxwriter==3.0.8
python-multipart==0.0.5
six==1.16.0
certifi==2022.12.7
charset-normalizer==3.0.1 
requests==2.28.2 
urllib3==1.26.14
et-xmlfile==1.1.0